Added tracking for number of skills being learned
Added the option to keep track of how many skills are 'learning'.  New Configuration variables (can be set by the GUI):
ExpTracker.CountSkills - Set to keep track (in a temp variable) and echo the number skills that are greater than a minimum mindstate.
ExpTracker.CountMinMindstate - The integer value of mindstate that is used to compare. (mindstate must be larger than this value to be counted.  Acceptable values 0-33)

Active variables:
(temporary) $ExpTracker.LearningSkills=32

Bug Fixes:
- N/A

Planned changes (still in wip):
- Support for Rested Exp in Exp Window
- Breaking out some of the config from code to XML files:
  * Support for modification of skills sorting/shortnames
  * (limited) support for custom reports